move cordova hooks into thali/install and get rid of `before_plugin_install.js`
