+++ plugin.xml
@@ -16,6 +16,8 @@
 
     <!-- Android -->
     <platform name="android">
+      <hook type="after_plugin_install" src="thali/install/cordova-hooks/android/after_plugin_install.js" />
+
         <config-file target="AndroidManifest.xml" parent="/*">
             <uses-permission android:name="android.permission.INTERNET" />
             <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -50,16 +52,14 @@
         </js-module>
         <source-file src="src/android/java/org/thaliproject/p2p/ThaliPermissions.java" target-dir="src/org/thaliproject/p2p/" />
 
-        <hook type="before_plugin_install" src="scripts/before_plugin_install.js" />
-        <hook type="before_compile" src="scripts/android/before_compile.js" />
     </platform>
 
     <!-- iOS -->
 
     <platform name="ios">
 
-        <hook type="before_plugin_install" src="scripts/before_plugin_install.js" />
-        <hook type="after_plugin_install" src="scripts/ios/after_plugin_install.js" />
+        <hook type="before_plugin_install" src="thali/install/cordova-hooks/ios/before_plugin_install.js" />
+        <hook type="after_plugin_install" src="thali/install/cordova-hooks/ios/after_plugin_install.js" />
 
         <config-file target="config.xml" parent="/*">
           <feature name="ThaliPlugin">
@@ -73,3 +73,4 @@
 
     </platform>
 </plugin>
+      <hook type="after_plugin_install" src="thali/install/cordova-hooks/android/after_plugin_install.js" />
