+++ thali/install/cordova-hooks/android/after_plugin_install.js
@@ -1,3 +1,9 @@
+//
+//  Copyright (C) Microsoft. All rights reserved.
+//  Licensed under the MIT license. See LICENSE.txt file in the project root
+//  for full license information.
+//
+
 'use strict';
 
 var fs = require('fs-extra-promise');
