+++ thali/install/cordova-hooks/package.json
